+++ resources/ts/layouts/components/batch-updaters/PayModuleBatchOverview.vue
@@ -166,10 +166,14 @@ const setPaymentTermMin = () => {
 
 const batchRemove = async () => {
     const count = props.selected_idxs.length
-    if (await alert.value.show('정말 ' + count + '개의 결제모듈을 일괄삭제 하시겠습니까?')) {
-        const params = { selected_idxs: props.selected_idxs }
-        const r = await request({ url: `/api/v1/manager/merchandises/pay-modules/batch-updaters/remove`, method: 'delete', data: params }, true)
-        emits('update:select_idxs', [])
+    if(count === 0)
+        snackbar.value.show('결제모듈을 1개이상 선택해주세요.', 'error')
+    else {
+        if (await alert.value.show('정말 ' + count + '개의 결제모듈을 일괄삭제 하시겠습니까?')) {
+            const params = { selected_idxs: props.selected_idxs }
+            const r = await request({ url: `/api/v1/manager/merchandises/pay-modules/batch-updaters/remove`, method: 'delete', data: params }, true)
+            emits('update:select_idxs', [])
+        }
     }
 }
