+++ resources/ts/layouts/components/batch-updaters/SalesBatchOverview.vue
@@ -104,10 +104,14 @@ const setNote = () => {
 
 const batchRemove = async () => {
     const count = props.selected_idxs.length
-    if (await alert.value.show('정말 ' + count + '개의 영업점을 일괄삭제 하시겠습니까?')) {
-        const params = { selected_idxs: props.selected_idxs }
-        const r = await request({ url: `/api/v1/manager/salesforces/batch-updaters/remove`, method: 'delete', data: params }, true)
-        emits('update:select_idxs', [])
+    if(count === 0)
+        snackbar.value.show('영업점을 1개이상 선택해주세요.', 'error')
+    else {
+        if (await alert.value.show('정말 ' + count + '개의 영업점을 일괄삭제 하시겠습니까?')) {
+            const params = { selected_idxs: props.selected_idxs }
+            const r = await request({ url: `/api/v1/manager/salesforces/batch-updaters/remove`, method: 'delete', data: params }, true)
+            emits('update:select_idxs', [])
+        }
     }
 }
